command for populating sampel data
